# src/SimpleTaskManager.Api/Models/NewBoardModel.cs
using System;
using System.Collections.Generic;
using System.Linq;
using System.Threading.Tasks;

namespace SimpleTaskManager.Api.Models
{
    public class NewBoardModel
    {
        public string Name { get; set; }
    }
}